Passing tests, prepped for other test cases. Still a bit of additional cleanup to do
