--- test/test_nodes.go
@@ -1,1 +1,0 @@
-package test
